[feat]#6 section1 구현
