+++ 2nd-Seminar/2nd-Seminar/UIView+.swift
@@ -1,0 +1,10 @@
+import UIKit
+
+extension UIView {
+        func addSubviews(_ views: UIView...) {
+        views.forEach {
+            self.addSubview($0)
+            $0.translatesAutoresizingMaskIntoConstraints = false
+        }
+    }
+}
